Binary tree function was updated to make the code a bit more consistent.
The tree now has 1 root node 4 right and 4 left subtree nodes.

The nodes are placed by offsetting a varying amount of X and Y.
This was done to limit collisions(where the tree overlaps on itself).
While collisions may still happen they are very rare.
//could implement a collision detection function and just generate a new tree on collision
Next step will be to split the function into 2 separate functions
 one to generate the array
 one for the actual placement of the tree

UI size settings were changed to accommodate the tree
The display pane was made larger as well as the surrounding text to match