Binary Tree update
The single function was broken down into 3 more basic components.
This allows for potential parsing of the information for animations later.

There is now a function to generate random tree values.
The reason there is a method for picking the random numbers is to ensure a more balanced tree.

There is now a function to generate a tree by passing in an array of your own integers.
This can be modified later to pass a set of arrays for animations.
This will also be needed for later input methods

There is now a function to draw the tree onto the pane.
This may be modified later to only go one step at a time if we want to include the tree construction
into the animation.